pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "purdue-company"
version = "0.1.0"
description = "CrewAI multi-agent company: Robotics R&D + Financial Trading"
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.80.0",
    "crewai-tools>=0.20.0",
    "anthropic>=0.40.0",
    "yfinance>=0.2.50",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
